// typechain-types/MempoolShield.ts
export interface MempoolShieldInterface extends Interface {
  getFunction(
    nameOrSignature: "commit" | "getCommit" | "minRevealDelay" | "reveal"
  ): FunctionFragment;

  getEvent(nameOrSignatureOrTopic: "Committed" | "Revealed"): EventFragment;

  encodeFunctionData(functionFragment: "commit", values: [BytesLike]): string;
  encodeFunctionData(
    functionFragment: "getCommit",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "minRevealDelay",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "reveal",
    values: [string, string]
  ): string;

  decodeFunctionResult(functionFragment: "commit", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "getCommit", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "minRevealDelay",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "reveal", data: BytesLike): Result;
}

export namespace CommittedEvent {
  export type InputTuple = [user: AddressLike, commitment: BytesLike];
  export type OutputTuple = [user: string, commitment: string];
  export interface OutputObject {
    user: string;
    commitment: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace RevealedEvent {
  export type InputTuple = [user: AddressLike, action: string];
  export type OutputTuple = [user: string, action: string];
  export interface OutputObject {
    user: string;
    action: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface MempoolShield extends BaseContract {
  connect(runner?: ContractRunner | null): MempoolShield;
  waitForDeployment(): Promise<this>;

  interface: MempoolShieldInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  commit: TypedContractMethod<[_commitment: BytesLike], [void], "nonpayable">;

  getCommit: TypedContractMethod<
    [user: AddressLike],
    [
      [string, bigint, boolean] & {
        commitment: string;
        blockNumber: bigint;
        revealed: boolean;
      }
    ],
    "view"
  >;

  minRevealDelay: TypedContractMethod<[], [bigint], "view">;

  reveal: TypedContractMethod<
    [action: string, salt: string],
    [void],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "commit"
  ): TypedContractMethod<[_commitment: BytesLike], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "getCommit"
  ): TypedContractMethod<
    [user: AddressLike],
    [
      [string, bigint, boolean] & {
        commitment: string;
        blockNumber: bigint;
        revealed: boolean;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "minRevealDelay"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "reveal"
  ): TypedContractMethod<[action: string, salt: string], [void], "nonpayable">;

  getEvent(
    key: "Committed"
  ): TypedContractEvent<
    CommittedEvent.InputTuple,
    CommittedEvent.OutputTuple,
    CommittedEvent.OutputObject
  >;
  getEvent(
    key: "Revealed"
  ): TypedContractEvent<
    RevealedEvent.InputTuple,
    RevealedEvent.OutputTuple,
    RevealedEvent.OutputObject
  >;

  filters: {
    "Committed(address,bytes32)": TypedContractEvent<
      CommittedEvent.InputTuple,
      CommittedEvent.OutputTuple,
      CommittedEvent.OutputObject
    >;
    Committed: TypedContractEvent<
      CommittedEvent.InputTuple,
      CommittedEvent.OutputTuple,
      CommittedEvent.OutputObject
    >;

    "Revealed(address,string)": TypedContractEvent<
      RevealedEvent.InputTuple,
      RevealedEvent.OutputTuple,
      RevealedEvent.OutputObject
    >;
    Revealed: TypedContractEvent<
      RevealedEvent.InputTuple,
      RevealedEvent.OutputTuple,
      RevealedEvent.OutputObject
    >;
  };
}